ADD : 로봇 상태 카드 공용 컴포넌트로 분리
