+++ pages/robot/index.tsx
@@ -8,6 +8,7 @@ import { storeNameState } from '@src/store/storeNameState';
 import robotAPI from '@src/api/robot';
 import { IRobotState } from '@src/types/robot';
 import { IRobotDataList } from '@src/components/common/Dropdown/types';
+import RobotState from '@src/components/common/RobotState/RobotState';
 
 export async function getServerSideProps() {
   const robots = await robotAPI.getRobots();
@@ -37,7 +38,6 @@ const Robot = ({ stores, robots }: IProps) => {
   const setStoreName = useSetRecoilState(storeNameState);
 
   const getRobotsByState = (state: string) => {
-    console.log(state);
     const filteredList = state === '0' ? robots.robot : robots.robot.filter(({ robot_state }) => robot_state === state);
     return setRobotList(filteredList);
   };
@@ -45,14 +45,6 @@ const Robot = ({ stores, robots }: IProps) => {
   const pageHandler = (storeName: string, storeId: string) => {
     setStoreName(storeName);
     router.push(`/robot/${storeId}`);
-  };
-
-  const getColorById = (id: string) => {
-    return ROBOT_STATE.filter((robot) => robot.id === id)[0].color;
-  };
-
-  const getStateById = (id: string) => {
-    return ROBOT_STATE.filter((robot) => robot.id === id)[0].state;
   };
 
   return (
@@ -63,32 +55,11 @@ const Robot = ({ stores, robots }: IProps) => {
       </StHeader>
       <StBody>
         <StRobotList>
-          {robotList.map(({ k_map_name, serving_count, distance, state, battery, serial_number }, idx) => (
-            <StRobotState key={idx}>
-              <StFlexBox>
-                <StStore>{k_map_name}</StStore>
-                <StGapBox>
-                  <StState color={getColorById(state)}>{getStateById(state)}</StState>
-                  <StImg src={`assets/images/robot/robot_state_${state}.png`} alt="로봇이미지" />
-                </StGapBox>
-              </StFlexBox>
-              <StSerialNumber>S/N. {serial_number}</StSerialNumber>
-              <StGapBox>
-                <StGray>서빙횟수</StGray>
-                <StBlack>{serving_count ? serving_count : 0} 회</StBlack>
-              </StGapBox>
-              <StGapBox>
-                <StGray>이동거리</StGray>
-                <StBlack>{distance ? distance : 0} m</StBlack>
-              </StGapBox>
-              <StGapBox>
-                <StBatteryBox>
-                  <StBatteryBar width={Number(battery) > 0 ? Number(battery) : 0} />
-                </StBatteryBox>
-                <StBatteryPercent>{battery}%</StBatteryPercent>
-              </StGapBox>
-            </StRobotState>
-          ))}
+          {robotList.length !== 0 ? (
+            robotList.map((robot, idx) => <RobotState {...robot} key={idx} ROBOT_STATE={ROBOT_STATE} />)
+          ) : (
+            <StNull>해당 상태의 로봇이 존재하지 않습니다 ;(</StNull>
+          )}
         </StRobotList>
       </StBody>
     </StRobot>
@@ -120,6 +91,7 @@ const StBody = styled.div`
 `;
 
 const StRobotList = styled.div`
+  position: relative;
   margin-top: 1vw;
   width: 100%;
   display: grid;
@@ -128,89 +100,13 @@ const StRobotList = styled.div`
   gap: 1vw;
 `;
 
-const StRobotState = styled.div`
-  padding: 1vw;
-  background: white;
-  border-radius: 5px;
-  box-shadow: rgba(99, 99, 99, 0.2) 0vw 0.1042vw 0.4167vw 0vw;
-`;
-
-const StFlexBox = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const StStore = styled.p`
-  font-size: 12px;
-  font-weight: 500;
-`;
-
-const StGapBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1vw;
-`;
-
-const StState = styled.p<{ color: string | undefined }>`
-  color: ${({ color }) => color};
-  font-size: 13px;
-  font-weight: 600;
-`;
-
-const StImg = styled.img`
-  width: 30px;
-  height: 30px;
-`;
-
-const StSerialNumber = styled.p`
-  margin: 0.2vw 0;
-  text-align: right;
-  font-size: 10px;
-`;
-
-const StGray = styled.p`
-  font-size: 12px;
-  color: ${({ theme }) => theme.color.gray600};
-`;
-
-const StBlack = styled.p`
-  font-size: 12px;
-  color: ${({ theme }) => theme.color.black};
-`;
-
-const StBatteryBox = styled.div`
-  margin-top: 0.5vw;
-  position: relative;
-  width: 100%;
-  height: 0.3646vw;
-  background: #e8e6f0;
-  border-radius: 20px;
-  overflow: hidden;
-`;
-
-const StBatteryBar = styled.span<{ width: number | undefined }>`
+const StNull = styled.p`
   position: absolute;
-  display: block;
-  top: 0;
-  left: 0;
-  width: ${({ width }) => `${width}%`};
-  background: #4f79f9;
-  border-left: 1px solid $subPrimary-color;
-  opacity: 0.7;
-  height: 100%;
-  transition: width 0.5s ease-in-out;
-`;
-
-const StBatteryPercent = styled.p`
-  margin-top: 0.5vw;
-  font-size: 12px;
-  color: ${({ theme }) => theme.color.black};
-`;
-
-const StFooter = styled.div`
-  width: 100%;
+  top: 50%;
+  left: 50%;
+  translate: -50%;
+  color: ${({ theme }) => theme.color.gray500};
+  font-size: 24px;
 `;
 
 export default Robot;
