+++ src/components/Home/AllStores/Store.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled';
 import Image from 'next/image';
 import React from 'react';
-import { IStore } from './types';
 
 interface IProps {
   store: IStore;
