--- src/store/Sample.tsx
@@ -1,8 +1,0 @@
-//하단 코드 지우고 작성 시작
-import React from 'react';
-
-const Sample = () => {
-  return <div></div>;
-};
-
-export default Sample;
